+++ src/redux/stores/indexer/indexerHelpers.ts
@@ -1,25 +1,10 @@
 import { Server, SuccessResponse } from '@airswap/libraries';
-import {
-  FullOrder,
-  IndexedOrder,
-  OrderResponse,
-} from '@airswap/types';
+import { FullOrder, OrderResponse } from '@airswap/types';
 import { SortField, SortOrder } from '@airswap/types/build/src/server';
 
 import { INDEXER_ORDER_RESPONSE_TIME_MS } from '../../../constants/indexer';
 import { OrderFilter } from '../../../entities/OrderFilter/OrderFilter';
-
-const isPromiseFulfilledResult = <T>(result: any): result is PromiseFulfilledResult<T> => result && result.status === 'fulfilled' && 'value' in result;
-
-const isOrderResponse = <T>(resource: any): resource is OrderResponse<T> => (resource && 'orders' in resource);
-
-const isSuccessResponse = (response: any): response is SuccessResponse => response
-  && typeof response.message === 'string'
-  && !('code' in response);
-
-const getUndefinedAfterTimeout = (time: number): Promise<undefined> => new Promise<undefined>((resolve) => {
-  setTimeout(() => resolve(undefined), time);
-});
+import { getUndefinedAfterTimeout, isPromiseFulfilledResult, isSuccessResponse } from '../../../helpers/indexers';
 
 export const getOrdersFromServer = async (server: Server, filter: OrderFilter): Promise<OrderResponse<FullOrder> | undefined> => {
   const defaultFilter: Partial<OrderFilter> = {
@@ -93,33 +78,3 @@ export const sendOrderToIndexers = async (
   return isSuccessResponse(response);
 };
 
-export const getOrdersFromIndexers = async (filter: OrderFilter, indexerUrls: string[]): Promise<FullOrder[]> => {
-  if (!indexerUrls.length) {
-    console.error('[getOrdersFromIndexers] No indexer urls provided');
-  }
-
-  const servers = await getServers(indexerUrls);
-
-  const responses = await Promise.all(
-    servers.map(server => Promise.race([
-      getOrdersFromServer(server, filter),
-      getUndefinedAfterTimeout(INDEXER_ORDER_RESPONSE_TIME_MS),
-    ])),
-  );
-
-  const orderResponses = responses.filter(isOrderResponse<FullOrder>);
-
-  if (orderResponses.length === 0) {
-    console.error('[getOrdersFromIndexers] No order responses received');
-
-    throw new Error('No order responses received');
-  }
-
-  const indexedOrders: Record<string, IndexedOrder<FullOrder>> = orderResponses
-    .reduce((total, orderResponse) => ({
-      ...total,
-      ...orderResponse.orders,
-    }), {});
-
-  return Object.values(indexedOrders).map(indexedOrder => indexedOrder.order);
-};
